Mention new container port on upgrade overview
The new app listening port setting is a bit difficult to find, it'd be great to mention the change on the upgrade overview instructions. Seeing quite a few discussions on Discord about their app failing to boot when upgrading and this is one of the most common issues coming up.